Add Reset Method and Long Press Command Support to ZoomView
This PR introduces two new features to the Plugin.Maui.ZoomView control:

Reset Method:
Adds a Reset() method to programmatically reset the zoom and position of the ZoomView to its initial state.

LongPressedCommand:
Adds a LongPressedCommand bindable property, allowing developers to execute a custom command when the ZoomView is long pressed.